pyproject.toml
[project]
name = "DinfoStash"
version = "0.1.0"
description = "A Developer Information Store that can store and manage devloper's portfolio data and also generate resumes."
authors = [
    {name = "Amulya Paritosh", email = "amulya1171@gmail.com"},
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "pylatex>=1.4.2",
    "pydantic[email]>=2.9.2",
    "pydantic-extra-types>=2.9.0",
    "phonenumbers>=8.13.47",
    "fastapi>=0.115.2",
    "uvicorn[standard]>=0.32.0",
    "firebase-admin>=6.5.0",
    "jinja2>=3.1.4",
    "google-cloud-datastore>=2.20.1",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
]

[tool.pdm.dev-dependencies]
test = [
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "faker>=30.6.0",
]

[tool.pdm]
distribution = true

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
dinfostash = "dinfostash.cli:app"
